requirements.txt
Django==5.0.2
PyJWT==2.8.0
redis==5.0.1
django-redis==5.4.0
Pillow==10.2.0
python-dotenv==1.0.1
django-tailwind==3.8.0
django-crispy-forms==2.1
crispy-tailwind==1.0.0 